+++ src/App.js
@@ -3,7 +3,6 @@ import Layouts from "./layouts/Layouts";
 import Main from "./pages/Main"
 import Question from "./pages/Question";
 import Complete from "./pages/Complete";
-import Record from "./common/record";
 import StartAudio from "./pages/StartAudio";
 import AudioPlayer from "./pages/AudioPlayer";
 
@@ -18,7 +17,6 @@ function App() {
           <Route path="/" element={<Main />}/>
           <Route path="/question" element={<Question />} />
           <Route path="/complete" element={<Complete />} />
-          <Route path="/record" element={<Record />} />
           <Route path="/audioplayer" element={<AudioPlayer />} />
         </Routes>
       </BrowserRouter>
